refactor: move classes and related libs to common and small fixes
