+++ common-files/classes/fq2.mjs
@@ -3,8 +3,9 @@ Class representing a complex modular number.  In particular, this is useful
 for dealing with f_q^2 field elements in the alt BN128 curve.
 */
 
+// eslint-disable-next-line import/no-extraneous-dependencies
 import config from 'config';
-import { modDivide, complexDivMod } from '../utils/curve-maths/modular-division.mjs';
+import { modDivide, complexDivMod } from '../utils/crypto/modular-division.mjs';
 
 const { BN128_PRIME_FIELD } = config;
