+++ nightfall-optimist/src/classes/index.mjs
@@ -3,8 +3,7 @@ import PublicInputs from 'common-files/classes/public-inputs.mjs';
 import Transaction from 'common-files/classes/transaction.mjs';
 
 import BlockError from './block-error.mjs';
-import Fq2 from './fq2.mjs';
 import TransactionError from './transaction-error.mjs';
 import VerificationKey from './verification-key.mjs';
 
-export { Proof, PublicInputs, Transaction, BlockError, Fq2, TransactionError, VerificationKey };
+export { Proof, PublicInputs, Transaction, BlockError, TransactionError, VerificationKey };
